PDENLPModels.jl Documentation

JSOPDESolver

Krylov.jl to solve linear PDEs

Krylov.jl to solve the linear systems in the Newton-loop to solve nonlinear PDEs

PDENLPModels.jl Documentation

Functions

Index

References

PDENLPModels Progress

Functions

PDENLPModels.EnergyFETerm — Type

FETerm modeling the objective function of the optimization problem.

\[\begin{equation} \int_{\Omega} f(y,u) d\Omega, \end{equation}\]

where Ω is described by:

trian :: Triangulation

quad :: CellQuadrature

Constructor:

EnergyFETerm(:: Function, :: Triangulation, :: CellQuadrature)

See also: MixedEnergyFETerm, NoFETerm, _obj_cell_integral, _obj_integral, compute\gradient_k!

PDENLPModels.GridapPDENLPModel — Type

PDENLPModels using Gridap.jl

https://github.com/gridap/Gridap.jl Cite: Badia, S., & Verdugo, F. (2020). Gridap: An extensible Finite Element toolbox in Julia. Journal of Open Source Software, 5(52), 2520.

Find functions (y,u): Y -> ℜⁿ x ℜⁿ and κ ∈ ℜⁿ satisfying

min ∫_Ω​ f(κ,y,u) dΩ​ s.t. y solution of a PDE(κ,u)=0 lcon <= c(κ,y,u) <= ucon lvar <= (κ,y,u) <= uvar

     ```math
     \begin{aligned}
     \min_{κ,y,u} \ & ∫_Ω​ f(κ,y,u) dΩ​ \\
     \mbox{ s.t. } & y \mbox{ solution of } PDE(κ,u)=0, \\
     & lcon <= c(κ,y,u) <= ucon, \\
     & lvar <= (κ,y,u)  <= uvar.
     \end{aligned}
     ```

The weak formulation is then: res((y,u),(v,q)) = ∫ v PDE(κ,y,u) + ∫ q c(κ,y,u)

where the unknown (y,u) is a MultiField see Tutorials 7 and 8 of Gridap.

The set Ω​ is represented here with trian and quad.

TODO: [ ] time evolution pde problems. [ ] Handle the case where g and H are given. [ ] Handle several terms in the objective function (via an FEOperator)? [ ] Be more explicit on the different types of FETerm in fromtermtoterms! [ ] Could we control the Dirichlet boundary condition? (like classical control of heat equations) [ ] Clean the tests. [ ] Missing: constraint ncon with numfreedofs(Xpde)?

Main constructor:

GridapPDENLPModel(:: NLPModelMeta, :: Counters, :: AbstractEnergyTerm, :: FESpace, :: Union{FESpace,Nothing}, :: FESpace, :: Union{FESpace,Nothing}, :: FESpace, :: Union{FESpace,Nothing}, :: Union{FEOperator, Nothing}, :: Int, :: Int, :: Int)

Additional constructors:

Unconstrained and no control

GridapPDENLPModel(x0, tnrj, Ypde, Xpde) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde) GridapPDENLPModel(f, trian, quad, Ypde, Xpde)

Bound constraints:

GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde, lvar, uvar) GridapPDENLPModel(f, trian, quad, Ypde, Xpde, lvar, uvar)

PDE-constrained:

GridapPDENLPModel(x0, tnrj, Ypde, Xpde, c) GridapPDENLPModel(Ypde, Xpde, c) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, c) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, c) GridapPDENLPModel(f, trian, quad, Ypde, Ycon, Xpde, Xcon, c) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, c, lcon, ucon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, c, lcon, ucon) GridapPDENLPModel(f, trian, quad, Ypde, Ycon, Xpde, Xcon, c, lcon, ucon)

PDE-constrained and bounds:

GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar, c) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c) GridapPDENLPModel(f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, lcon, ucon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, lcon, ucon) GridapPDENLPModel(f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, lcon, ucon)

Future constructors:

Functional bounds: in this case |lvar|=|uvar|=nparam

GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lfunc, ufunc) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde, lfunc, ufunc) GridapPDENLPModel(f, trian, quad, Ypde, Xpde, lfunc, ufunc) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar, lfunc, ufunc) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde, lvar, uvar, lfunc, ufunc) GridapPDENLPModel(f, trian, quad, Ypde, Xpde, lvar, uvar, lfunc, ufunc) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar, lfunc, ufunc, c) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c) GridapPDENLPModel(f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, lcon, ucon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, lcon, ucon) GridapPDENLPModel(f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, lcon, ucon)

Discrete constraints (ck, lckon, uckon) only for problems with nparam > 0 (hence only if x0 given or tnrj)

GridapPDENLPModel(x0, tnrj, Ypde, Xpde, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, c, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, c, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, c, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, c, lcon, ucon, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, c, lcon, ucon, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar, c, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, lcon, ucon, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, c, lcon, ucon, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lfunc, ufunc, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde, lfunc, ufunc, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar, lfunc, ufunc, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Xpde, lvar, uvar, lfunc, ufunc, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Xpde, lvar, uvar, lfunc, ufunc, c, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, ck, lckon, uckon) GridapPDENLPModel(x0, tnrj, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, lcon, ucon, ck, lckon, uckon) GridapPDENLPModel(x0, f, trian, quad, Ypde, Ycon, Xpde, Xcon, lvar, uvar, lfunc, ufunc, c, lcon, ucon, ck, lckon, uckon)

The following keyword arguments are available to all constructors:

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints

(default: Int[] or 1:ncon if c is an AffineFEOperator)

The following keyword arguments are available to the constructors for constrained problems explictly giving lcon and ucon:

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

Notes:

We handle two types of FEOperator: AffineFEOperator, and FEOperatorFromTerms

which is the obtained by the FEOperator constructor. The terms supported in FEOperatorFromTerms are: FESource, NonlinearFETerm, NonlinearFETermWithAutodiff, LinearFETerm, AffineFETerm.

If lcon and ucon are not given, they are assumed zeros.

If the type can't be deduced from the argument, it is Float64.

PDENLPModels.MixedEnergyFETerm — Type

FETerm modeling the objective function of the optimization problem with functional and discrete unknowns.

\[\begin{equation} \int_{\Omega} f(y,u,\kappa) d\Omega, \end{equation}\]

where Ω is described by:

trian :: Triangulation

quad :: CellQuadrature

Constructor:

MixedEnergyFETerm(:: Function, :: Triangulation, :: CellQuadrature, :: Int)

See also: EnergyFETerm, NoFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.NoFETerm — Type

FETerm modeling the objective function when there are no intregral objective.

math \begin{equation} f(\kappa) \end{equation}

Constructors:

NoFETerm()

NoFETerm(:: Function)

See also: MixedEnergyFETerm, EnergyFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.ResidualEnergyFETerm — Type

FETerm modeling the objective function of the optimization problem with functional and discrete unknowns, describe as a norm and a regularizer.

\[\begin{equation} \frac{1}{2}\|Fyu(y,u)\|^2_{L^2_\Omega} + \lambda\int_{\Omega} lyu(y,u) d\Omega + \frac{1}{2}\|Fk(κ)\|^2 + \mu lk(κ) \end{equation}\]

where Ω is described by:

trian :: Triangulation

quad :: CellQuadrature

Constructor:

ResidualEnergyFETerm(:: Function, :: Triangulation, :: CellQuadrature, :: Function, :: Int)

See also: EnergyFETerm, NoFETerm, MixedEnergyFETerm

PDENLPModels.VoidFESpace — Type

Useful structure to avoid putting nothing

NLPModels.jprod! — Method

Jv = jprod!(nlp, x, v, Jv)

Evaluate $J(x)v$, the Jacobian-vector product at x in place.

Note for GridapPDENLPModel:

Evaluate the jacobian and then use mul! (here coo_prod! is slower as we have to compute findnz).

Alternative: benefit from the AD? Jv .= ForwardDiff.derivative(t->nlp.c(nlp, x + t * v), 0)

when the jacobian is obtained by AD.

NLPModels.jtprod! — Method

Jv = jtprod!(nlp, x, v, Jv)

Evaluate $J(x)'v$, the Jacobian-vector product at x in place.

Note for GridapPDENLPModel:

Evaluate the jacobian and then use mul! (here coo_prod! is slower as we have to compute findnz).

Alternative: benefit from the AD? Jtv .= ForwardDiff.gradient(x -> dot(nlp.c(x), v), x)

when the jacobian is obtained by AD.

PDENLPModels._compute_gradient! — Function

Return the gradient of the objective function and set it in place.

_compute_gradient!(:: AbstractVector, :: EnergyFETerm, :: AbstractVector, :: FEFunctionType, :: FESpace, :: FESpace)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_hess_coo, _compute_hess_k_coo

PDENLPModels._compute_gradient_k — Function

Return the derivative of the objective function w.r.t. κ.

_compute_gradient_k(:: AbstractEnergyTerm, :: FEFunctionType, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_hess_coo, _compute_hess_k_coo

PDENLPModels._compute_hess_coo — Function

Return the hessian w.r.t. yu of the objective function in coo format.

_compute_hess_coo(:: AbstractEnergyTerm, :: AbstractVector, :: FEFunctionType, :: FESpace, :: FESpace)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_gradient_k, _compute_hess_k_coo

PDENLPModels._compute_hess_k_coo — Function

Return the hessian w.r.t. κ of the objective function in coo format.

_compute_hess_k_coo(:: AbstractNLPModel, :: AbstractEnergyTerm, :: AbstractVector, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_gradient_k, _compute_hess_coo

PDENLPModels._compute_hess_k_vals — Function

Return the values of the hessian w.r.t. κ of the objective function.

_compute_hess_k_vals(:: AbstractNLPModel, :: AbstractEnergyTerm, :: AbstractVector, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_gradient_k, _compute_hess_coo

PDENLPModels._fill_hess_at_cell! — Method

https://github.com/gridap/Gridap.jl/blob/758a8620756e164ba0e6b83dc8dcbb278015b3d9/src/FESpaces/SparseMatrixAssemblers.jl#L463 fillmatrixatcell! may have a specific specialization

PDENLPModels._from_terms_to_jacobian — Method

Note:

Compute the derivatives w.r.t. y and u separately.

Use AD for those derivatives. Only for the following:

NonlinearFETerm (we neglect the inapropriate jac function);

NonlinearFETermWithAutodiff

TODO: Gridap.FESpaces.FETerm & AffineFETerm ?

FESource <: AffineFETerm (jacobian of a FESource is nothing)

LinearFETerm <: AffineFETerm (not implemented)

PDENLPModels._from_terms_to_jacobian2 — Method

Would be better but somehow autodiffcelljacobianfromresidual is restricted to square matrices at some point.

PDENLPModels._from_terms_to_residual! — Method

Note:

mul! seems faster than doing:

rows, cols, vals = findnz(getmatrix(op)) cooprod!(cols, rows, vals, v, res)

get_matrix(op) is a sparse matrix

Benchmark equivalent to Gridap.FESpaces.residual!(res, op_affine.op, xrand)

PDENLPModels._functions_to_vectors! — Method

functionstovectors!(nini :: Int, nfields :: Int, trian :: Triangulation, lfunc :: Function, ufunc :: Function, cellxm :: Gridap.Arrays.AppliedArray, Y :: FESpace, lvar :: AbstractVector, uvar :: AbstractVector)

Iterate for k = 1 to nfields and switch lfunc[k] and ufunc[k] to vectors, allocated in lvar and uvar in place starting from nini + 1. It returns nini + the number of allocations.

PDENLPModels._jac_structure! — Method

Adaptation of function allocate_matrix(a::SparseMatrixAssembler,matdata) end in Gridap.FESpaces.

PDENLPModels._obj_cell_integral — Function

Return the integral of the objective function

_obj_cell_integral(:: AbstractEnergyTerm, :: GenericCellField, :: AbstractVector)

x is a vector of GenericCellField, for instance resulting from yuh = CellField(Y, cell_yu).

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _compute_gradient_k, _compute_hess_coo, _compute_hess_k_coo

PDENLPModels._obj_integral — Function

Return the integral of the objective function

_obj_integral(:: AbstractEnergyTerm, :: FEFunctionType, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_cell_integral, _compute_gradient_k, _compute_hess_coo, _compute_hess_k_coo

PDENLPModels._split_FEFunction — Method

_split_FEFunction(:: AbstractVector, :: FESpace, :: Union{FESpace, Nothing})

Split the vector x into two FEFunction corresponding to the solution y and the control u. Returns nothing for the control u if Ycon == nothing.

Do not verify the compatible length.

PDENLPModels._split_vector — Method

_split_vector(:: AbstractVector, :: FESpace, :: Union{FESpace, Nothing})

Split the vector x into three vectors: y, u, k. Returns nothing for the control u if Ycon == nothing.

Do not verify the compatible length.

PDENLPModels._struct_hess_at_cell! — Method

https://github.com/gridap/Gridap.jl/blob/758a8620756e164ba0e6b83dc8dcbb278015b3d9/src/FESpaces/SparseMatrixAssemblers.jl#L463 fillmatrixatcell! may have a specific specialization

PDENLPModels._vals_hess_at_cell! — Method

https://github.com/gridap/Gridap.jl/blob/758a8620756e164ba0e6b83dc8dcbb278015b3d9/src/FESpaces/SparseMatrixAssemblers.jl#L463 fillmatrixatcell! may have a specific specialization

PDENLPModels.assemble_hess — Method

These functions: https://github.com/gridap/Gridap.jl/blob/758a8620756e164ba0e6b83dc8dcbb278015b3d9/src/FESpaces/SparseMatrixAssemblers.jl#L463

https://github.com/gridap/Gridap.jl/blob/758a8620756e164ba0e6b83dc8dcbb278015b3d9/src/Algebra/SparseMatrixCSC.jl

https://github.com/gridap/Gridap.jl/blob/758a8620756e164ba0e6b83dc8dcbb278015b3d9/src/Algebra/SparseMatrices.jl#L29-L33

PDENLPModels.assemble_jac_vals! — Method

Adaptation of function assemble_matrix_add!(mat,a::GenericSparseMatrixAssembler,matdata) end from Gridap.FESpaces

PDENLPModels.bounds_functions_to_vectors — Method

(lvar, uvar) = boundsfunctionsto_vectors(Y :: MultiFieldFESpace, Ycon :: Union{FESpace, Nothing}, Ypde :: FESpace, trian :: Triangulation, lyfunc :: Union{Function, AbstractVector}, uyfunc :: Union{Function, AbstractVector}, lufunc :: Union{Function, AbstractVector}, uufunc :: Union{Function, AbstractVector})

Return the bounds lvar and uvar.

PDENLPModels.count_nnz_coo_short — Method

This is a modified version of function count_matrix_nnz_coo(a::GenericSparseMatrixAssembler,matdata) end from Gridap.FESpaces. The motivation is to avoid the use of the yet unknown values.

PDENLPModels.fill_jac_coo_symbolic! — Method

This is a modified version of function fill_matrix_coo_symbolic!(I,J,a::GenericSparseMatrixAssembler,matdata,n=0) end from Gridap.FESpaces. The motivation is to avoid the use of the yet unknown values.

PDENLPModels.get_nnzh — Method

get_nnzh: return the number of non-zeros elements in the hessian matrix.

Different variants:

get_nnzh(tnrj :: T, Y, X, nparam, nvar): consider the hessian of the

objective function only.

get_nnzh(tnrj :: T, op :: AffineFEOperator, Y, X, nparam, nvar): consider

the hessian of the objective function only.

get_nnzh(tnrj :: T, op :: Gridap.FESpaces.FEOperatorFromTerms, Y, X, nparam, nvar):

concatenate non-zeros of the objective-hessian and the hessian of each term composing op.

TODO: Do not handle non-linear discrete parameters in the constraints.

PDENLPModels.hess_coo — Method

hess_coo: return the hessian of the constraints in COO-format.

Notes:

hess_coo(nlp, op :: AffineFEOperator, x, λ): return 0-matrix

hess_coo(nlp, op :: FEOperatorFromTerms, x, λ): iterate over the terms

TODO: make it a real COO-format function.

do not work with parameters.

PDENLPModels.hess_obj_structure — Method

hess_structure returns the sparsity pattern of the Lagrangian Hessian in sparse coordinate format, and hess_obj_structure is only for the objective function hessian.

PDENLPModels.hess_old — Method

julia> @btime hess(nlp, sol_gridap); 614.938 ms (724536 allocations: 90.46 MiB)

julia> @btime hessold(nlp, solgridap); 643.689 ms (724599 allocations: 127.48 MiB)

PDENLPModels.hessian_lagrangian_test_functions — Method

Function testing all the Lagrangian Hessian-related function implemented for GridapPDENLPModel.

return true if the test passed.

set udc to true to use NLPModels derivative check (can be slow)

https://github.com/JuliaSmoothOptimizers/NLPModels.jl/blob/master/src/dercheck.jl

should be used for small problems only as it computes hessians several times.

We test the functions from hessian_func.jl and normal test functions.

List of functions we are testing:

hess_coo ✓

hess ✓

hess_obj_structure ✓

hess_coord and hess_coord! ✓

hprod ✓

hess_op and hess_op! ✓

PDENLPModels.hessian_test_functions — Method

Function testing all the hessian-related function implemented for GridapPDENLPModel.

return true if the test passed.

set udc to true to use NLPModels derivative check (can be slow)

https://github.com/JuliaSmoothOptimizers/NLPModels.jl/blob/master/src/dercheck.jl

should be used for small problems only as it computes hessians several times.

We test the functions from hessian_func.jl and normal test functions.

List of functions we are testing:

hess_coo ✓

hess ✓

hess_obj_structure and hess_obj_structure! ✓

hess_coord and hess_coord! ✓

count_hess_nnz_coo_short ✓

hprod ✓

hess_op and hess_op! ✓

Comments:

set the number of "nnzh" (with the repeated entries?) in the META? Note, this is

necessary to have hess_structure! and hess_coord!.

count_hess_nnz_coo is used in the computation of hess_coo instead of

count_hess_nnz_coo_short, so if there is a difference it would appear in the tests (implicit test).

Use sparse differentiation

PDENLPModels.hprod_autodiff! — Method

Compute hessian-vector product of the objective function.

Note: this is not efficient at all. Test on n=14115 @btime hprod(nlp, solgridap, v) 42.683 s (274375613 allocations: 29.89 GiB) while computing the hessian and then the product yields @btime _Hx = hess(nlp, solgridap); 766.036 ms (724829 allocations: 121.84 MiB) @btime hprod(nlp, sol_gridap, v) 42.683 s (274375613 allocations: 29.89 GiB)

PDENLPModels.jacobian_test_functions — Method

Function testing all the jacobian-related function implemented for GridapPDENLPModel.

return true if the test passed.

set udc to true to use NLPModels derivative check (can be slow)

https://github.com/JuliaSmoothOptimizers/NLPModels.jl/blob/master/src/dercheck.jl

should be used for small problems only as it computes hessians several times.

We test the functions from jacobian_func.jl and normal test functions.

List of functions we are testing:

jac ✓

jac_structure and jac_structure! ✓

jac_coord and jac_coord! ✓

count_nnz_jac ✓

jprod and jtprod

jac_op and jac_op! ✓

Comments:

set the number of "nnzh" (with the repeated entries?) in the META?

Use sparse differentiation

Index

PDENLPModels.EnergyFETerm

PDENLPModels.GridapPDENLPModel

PDENLPModels.MixedEnergyFETerm

PDENLPModels.NoFETerm

PDENLPModels.ResidualEnergyFETerm

PDENLPModels.VoidFESpace

NLPModels.jprod!

NLPModels.jtprod!

PDENLPModels._compute_gradient!

PDENLPModels._compute_gradient_k

PDENLPModels._compute_hess_coo

PDENLPModels._compute_hess_k_coo

PDENLPModels._compute_hess_k_vals

PDENLPModels._fill_hess_at_cell!

PDENLPModels._from_terms_to_jacobian

PDENLPModels._from_terms_to_jacobian2

PDENLPModels._from_terms_to_residual!

PDENLPModels._functions_to_vectors!

PDENLPModels._jac_structure!

PDENLPModels._obj_cell_integral

PDENLPModels._obj_integral

PDENLPModels._split_FEFunction

PDENLPModels._split_vector

PDENLPModels._struct_hess_at_cell!

PDENLPModels._vals_hess_at_cell!

PDENLPModels.assemble_hess

PDENLPModels.assemble_jac_vals!

PDENLPModels.bounds_functions_to_vectors

PDENLPModels.count_nnz_coo_short

PDENLPModels.fill_jac_coo_symbolic!

PDENLPModels.get_nnzh

PDENLPModels.hess_coo

PDENLPModels.hess_obj_structure

PDENLPModels.hess_old

PDENLPModels.hessian_lagrangian_test_functions

PDENLPModels.hessian_test_functions

PDENLPModels.hprod_autodiff!

PDENLPModels.jacobian_test_functions

References

https://github.com/gridap/Gridap.jl Badia, S., & Verdugo, F. (2020). Gridap: An extensible Finite Element toolbox in Julia. Journal of Open Source Software, 5(52), 2520.

https://github.com/JuliaSmoothOptimizers/NLPModels.jl D. Orban and A. S. Siqueira and {contributors} (2020). NLPModels.jl: Data Structures for Optimization Models