Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PDENLPModels.EnergyFETerm

PDENLPModels.GridapPDENLPModel

PDENLPModels.MixedEnergyFETerm

PDENLPModels.NoFETerm

PDENLPModels.PDENLPMeta

PDENLPModels._compute_gradient!

PDENLPModels._compute_gradient_k!

PDENLPModels._compute_hess_k_vals!

PDENLPModels._compute_hess_structure

PDENLPModels._functions_to_vectors!

PDENLPModels._obj_integral

PDENLPModels._split_FEFunction

PDENLPModels._split_vector

PDENLPModels._split_vectors

PDENLPModels.bounds_functions_to_vectors

PDENLPModels.get_Hcols

PDENLPModels.get_Hrows

PDENLPModels.get_Jcols

PDENLPModels.get_Jrows

PDENLPModels.get_X

PDENLPModels.get_Xcon

PDENLPModels.get_Xpde

PDENLPModels.get_Y

PDENLPModels.get_Ycon

PDENLPModels.get_Ypde

PDENLPModels.get_nnzh_obj

PDENLPModels.get_nparam

PDENLPModels.get_nvar_con

PDENLPModels.get_nvar_pde

PDENLPModels.get_op

PDENLPModels.get_tnrj

PDENLPModels.split_vectors

​

PDENLPModels.EnergyFETerm — Type

FETerm modeling the objective function of the optimization problem.

\[\begin{aligned} \int_{\Omega} f(y,u) d\Omega, \end{aligned}\]

Constructor:

EnergyFETerm(:: Function, :: Triangulation, :: Measure)

See also: MixedEnergyFETerm, NoFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.GridapPDENLPModel — Type

PDENLPModels using Gridap.jl

https://github.com/gridap/Gridap.jl Cite: Badia, S., & Verdugo, F. (2020). Gridap: An extensible Finite Element toolbox in Julia. Journal of Open Source Software, 5(52), 2520.

Find functions (y,u): Y -> ℜⁿ x ℜⁿ and κ ∈ ℜⁿ satisfying

min ∫_Ω​ f(κ,y,u) dΩ​ s.t. y solution of a PDE(κ,u)=0 lcon <= c(κ,y,u) <= ucon lvar <= (κ,y,u) <= uvar

     ```math
     \begin{aligned}
     \min_{κ,y,u} \ & ∫_Ω​ f(κ,y,u) dΩ​ \\
     \mbox{ s.t. } & y \mbox{ solution of } PDE(κ,u)=0, \\
     & lcon <= c(κ,y,u) <= ucon, \\
     & lvar <= (κ,y,u)  <= uvar.
     \end{aligned}
     ```

The weak formulation is then: res((y,u),(v,q)) = ∫ v PDE(κ,y,u) + ∫ q c(κ,y,u)

where the unknown (y,u) is a MultiField see Tutorials 7 and 8 of Gridap.

Main constructor:

GridapPDENLPModel(:: NLPModelMeta, :: Counters, :: AbstractEnergyTerm, :: FESpace, :: Union{FESpace,Nothing}, :: FESpace, :: Union{FESpace,Nothing}, :: FESpace, :: Union{FESpace,Nothing}, :: Union{FEOperator, Nothing}, :: Int, :: Int, :: Int)

The following keyword arguments are available to all constructors:

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints

(default: Int[] or 1:ncon if c is an AffineFEOperator)

The following keyword arguments are available to the constructors for constrained problems explictly giving lcon and ucon:

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

Notes:

We handle two types of FEOperator: AffineFEOperator, and FEOperatorFromWeakForm

If lcon and ucon are not given, they are assumed zeros.

If the type can't be deduced from the argument, it is Float64.

PDENLPModels.MixedEnergyFETerm — Type

FETerm modeling the objective function of the optimization problem with functional and discrete unknowns.

\[\begin{aligned} \int_{\Omega} f(y,u,\kappa) d\Omega, \end{aligned}\]

Constructor:

MixedEnergyFETerm(:: Function, :: Triangulation, :: Int)

See also: EnergyFETerm, NoFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.NoFETerm — Type

FETerm modeling the objective function when there are no integral objective.

math \begin{aligned} f(\kappa) \end{aligned}

Constructors:

NoFETerm()

NoFETerm(:: Function)

See also: MixedEnergyFETerm, EnergyFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.PDENLPMeta — Type

PDENLPMeta

A composite type that represents the main features of the PDE-constrained optimization problem

The following arguments are accepted:

tnrj : structure representing the objective term

Ypde : TrialFESpace for the solution of the PDE

Ycon : TrialFESpace for the parameter

Xpde : TestFESpace for the solution of the PDE

Xcon : TestFESpace for the parameter

Y : concatenated TrialFESpace

X : concatenated TestFESpace

op : operator representing the PDE-constraint (nothing if no constraints)

nvar_pde :number of dofs in the solution functions

nvar_con : number of dofs in the control functions

nparam: : number of real unknowns

nnzh_obj : number of nonzeros elements in the objective hessian

Hrows : store the structure for the hessian of the lagrangian

Hcols : store the structure for the hessian of the lagrangian

Jrows : store the structure for the hessian of the jacobian

Jcols : store the structure for the hessian of the jacobian

PDENLPModels._compute_gradient! — Function

Return the gradient of the objective function and set it in place.

_compute_gradient!(:: AbstractVector, :: EnergyFETerm, :: AbstractVector, :: FEFunctionType, :: FESpace, :: FESpace)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_hess_k_coo

PDENLPModels._compute_gradient_k! — Function

Return the derivative of the objective function w.r.t. κ.

_compute_gradient_k!(::AbstractVector, :: AbstractEnergyTerm, :: FEFunctionType, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_hess_k_coo

PDENLPModels._compute_hess_k_vals! — Function

Return the values of the hessian w.r.t. κ of the objective function.

_compute_hess_k_vals!(:: AbstractVector, :: AbstractNLPModel, :: AbstractEnergyTerm, :: AbstractVector, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_integral, _obj_cell_integral, _compute_gradient_k

PDENLPModels._compute_hess_structure — Method

Return a triplet with the structure (rows, cols) and the number of non-zeros elements in the hessian w.r.t. y and u of the objective function.

The rows and cols returned by computehessstructureobj are already shifter by nparam.

PDENLPModels._functions_to_vectors! — Method

functionstovectors!(nini :: Int, nfields :: Int, trian :: Triangulation, lfunc :: Function, ufunc :: Function, cellxm :: Gridap.Arrays.AppliedArray, Y :: FESpace, lvar :: AbstractVector, uvar :: AbstractVector)

Iterate for k = 1 to nfields and switch lfunc[k] and ufunc[k] to vectors, allocated in lvar and uvar in place starting from nini + 1. It returns nini + the number of allocations.

PDENLPModels._obj_integral — Function

Return the integral of the objective function

_obj_integral(:: AbstractEnergyTerm, :: FEFunctionType, :: AbstractVector)

See also: MixedEnergyFETerm, EnergyFETerm, NoFETerm, _obj_cell_integral, _compute_gradient_k, _compute_hess_k_coo

PDENLPModels._split_FEFunction — Method

_split_FEFunction(:: AbstractVector, :: FESpace, :: Union{FESpace, Nothing})

Split the vector x into two FEFunction corresponding to the solution y and the control u. Returns nothing for the control u if Ycon == nothing.

Do not verify the compatible length.

PDENLPModels._split_vector — Method

_split_vector(:: AbstractVector, :: FESpace, :: Union{FESpace, Nothing})

Split the vector x into three vectors: y, u, k. Returns nothing for the control u if Ycon == nothing.

Do not verify the compatible length.

PDENLPModels._split_vectors — Method

`_split_vectors(x, Ypde, Ycon)`

Take a vector x and returns a splitting in terms of y, u and θ.

PDENLPModels.bounds_functions_to_vectors — Method

(lvar, uvar) = bounds_functions_to_vectors(Y :: MultiFieldFESpace, Ycon :: Union{FESpace, Nothing}, Ypde :: FESpace, trian :: Triangulation, lyfunc :: Union{Function, AbstractVector}, uyfunc :: Union{Function, AbstractVector}, lufunc :: Union{Function, AbstractVector}, uufunc :: Union{Function, AbstractVector})

Return the bounds lvar and uvar.

PDENLPModels.get_Hcols — Method

get_Hcols(nlp)
get_Hcols(meta)

Return the value Hcols from meta or nlp.meta.

PDENLPModels.get_Hrows — Method

get_Hrows(nlp)
get_Hrows(meta)

Return the value Hrows from meta or nlp.meta.

PDENLPModels.get_Jcols — Method

get_Jcols(nlp)
get_Jcols(meta)

Return the value Jcols from meta or nlp.meta.

PDENLPModels.get_Jrows — Method

get_Jrows(nlp)
get_Jrows(meta)

Return the value Jrows from meta or nlp.meta.

PDENLPModels.get_X — Method

get_X(nlp)
get_X(meta)

Return the value X from meta or nlp.meta.

PDENLPModels.get_Xcon — Method

get_Xcon(nlp)
get_Xcon(meta)

Return the value Xcon from meta or nlp.meta.

PDENLPModels.get_Xpde — Method

get_Xpde(nlp)
get_Xpde(meta)

Return the value Xpde from meta or nlp.meta.

PDENLPModels.get_Y — Method

get_Y(nlp)
get_Y(meta)

Return the value Y from meta or nlp.meta.

PDENLPModels.get_Ycon — Method

get_Ycon(nlp)
get_Ycon(meta)

Return the value Ycon from meta or nlp.meta.

PDENLPModels.get_Ypde — Method

get_Ypde(nlp)
get_Ypde(meta)

Return the value Ypde from meta or nlp.meta.

PDENLPModels.get_nnzh_obj — Method

get_nnzh_obj(nlp)
get_nnzh_obj(meta)

Return the value nnzh_obj from meta or nlp.meta.

PDENLPModels.get_nparam — Method

get_nparam(nlp)
get_nparam(meta)

Return the value nparam from meta or nlp.meta.

PDENLPModels.get_nvar_con — Method

get_nvar_con(nlp)
get_nvar_con(meta)

Return the value nvar_con from meta or nlp.meta.

PDENLPModels.get_nvar_pde — Method

get_nvar_pde(nlp)
get_nvar_pde(meta)

Return the value nvar_pde from meta or nlp.meta.

PDENLPModels.get_op — Method

get_op(nlp)
get_op(meta)

Return the value op from meta or nlp.meta.

PDENLPModels.get_tnrj — Method

get_tnrj(nlp)
get_tnrj(meta)

Return the value tnrj from meta or nlp.meta.

PDENLPModels.split_vectors — Method

`split_vectors(::GridapPDENLPModel, x)`

Take a vector x and returns a splitting in terms of y, u and θ.