Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PDENLPModels.EnergyFETerm

PDENLPModels.GridapPDENLPModel

PDENLPModels.MixedEnergyFETerm

PDENLPModels.NoFETerm

PDENLPModels.bounds_functions_to_vectors

​

PDENLPModels.EnergyFETerm — Type

FETerm modeling the objective function of the optimization problem.

\[\begin{aligned} \int_{\Omega} f(y,u) d\Omega, \end{aligned}\]

where Ω is described by:

trian :: Triangulation

quad :: Measure

Constructor:

EnergyFETerm(:: Function, :: Triangulation, :: Measure)

See also: MixedEnergyFETerm, NoFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.GridapPDENLPModel — Type

PDENLPModels using Gridap.jl

https://github.com/gridap/Gridap.jl Cite: Badia, S., & Verdugo, F. (2020). Gridap: An extensible Finite Element toolbox in Julia. Journal of Open Source Software, 5(52), 2520.

Find functions (y,u): Y -> ℜⁿ x ℜⁿ and κ ∈ ℜⁿ satisfying

min ∫_Ω​ f(κ,y,u) dΩ​ s.t. y solution of a PDE(κ,u)=0 lcon <= c(κ,y,u) <= ucon lvar <= (κ,y,u) <= uvar

     ```math
     \begin{aligned}
     \min_{κ,y,u} \ & ∫_Ω​ f(κ,y,u) dΩ​ \\
     \mbox{ s.t. } & y \mbox{ solution of } PDE(κ,u)=0, \\
     & lcon <= c(κ,y,u) <= ucon, \\
     & lvar <= (κ,y,u)  <= uvar.
     \end{aligned}
     ```

The weak formulation is then: res((y,u),(v,q)) = ∫ v PDE(κ,y,u) + ∫ q c(κ,y,u)

where the unknown (y,u) is a MultiField see Tutorials 7 and 8 of Gridap.

The set Ω​ is represented here with trian and quad.

Main constructor:

GridapPDENLPModel(:: NLPModelMeta, :: Counters, :: AbstractEnergyTerm, :: FESpace, :: Union{FESpace,Nothing}, :: FESpace, :: Union{FESpace,Nothing}, :: FESpace, :: Union{FESpace,Nothing}, :: Union{FEOperator, Nothing}, :: Int, :: Int, :: Int)

The following keyword arguments are available to all constructors:

name: The name of the model (default: "Generic")

The following keyword arguments are available to the constructors for constrained problems:

lin: An array of indexes of the linear constraints

(default: Int[] or 1:ncon if c is an AffineFEOperator)

The following keyword arguments are available to the constructors for constrained problems explictly giving lcon and ucon:

y0: An inital estimate to the Lagrangian multipliers (default: zeros)

Notes:

We handle two types of FEOperator: AffineFEOperator, and FEOperatorFromWeakForm

If lcon and ucon are not given, they are assumed zeros.

If the type can't be deduced from the argument, it is Float64.

PDENLPModels.MixedEnergyFETerm — Type

FETerm modeling the objective function of the optimization problem with functional and discrete unknowns.

\[\begin{aligned} \int_{\Omega} f(y,u,\kappa) d\Omega, \end{aligned}\]

where Ω is described by:

trian :: Triangulation

quad :: Measure

Constructor:

MixedEnergyFETerm(:: Function, :: Triangulation, :: Measure, :: Int)

See also: EnergyFETerm, NoFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.NoFETerm — Type

FETerm modeling the objective function when there are no integral objective.

math \begin{aligned} f(\kappa) \end{aligned}

Constructors:

NoFETerm()

NoFETerm(:: Function)

See also: MixedEnergyFETerm, EnergyFETerm, _obj_cell_integral, _obj_integral, _compute_gradient_k!

PDENLPModels.bounds_functions_to_vectors — Method

(lvar, uvar) = bounds_functions_to_vectors(Y :: MultiFieldFESpace, Ycon :: Union{FESpace, Nothing}, Ypde :: FESpace, trian :: Triangulation, lyfunc :: Union{Function, AbstractVector}, uyfunc :: Union{Function, AbstractVector}, lufunc :: Union{Function, AbstractVector}, uufunc :: Union{Function, AbstractVector})

Return the bounds lvar and uvar.